Fix for multichain network

+++ scripts/init-multichain.sh
@@ -1,7 +1,14 @@
 #!/usr/bin/env bash
 
+if ! type "hermes" > /dev/null; then
+  # install foobar here
+  echo "You need the hermes relayer to run this script. You can find it here https://github.com/informalsystems/ibc-rs"
+  exit 0
+fi
+
 ### chain init script for development purposes only ###
 killall sifnoded
+killall hermes
 rm -rf ~/.sifnode-1
 rm -rf ~/.sifnode-2
 rm -rf ~/.sifnode-3
@@ -97,38 +104,39 @@ sifnoded collect-gentxs --home ~/.sifnode-3
 echo "Validating genesis file..."
 sifnoded validate-genesis --home ~/.sifnode-3
 
+rm -rf abci_*.log
+rm -rf hermes.log
+rm -rf ~/.hermes
 
-
-
-sleep 1
-sifnoded start --home ~/.sifnode-1 --p2p.laddr 0.0.0.0:27655  --grpc.address 0.0.0.0:9090 --address tcp://0.0.0.0:27659 --rpc.laddr tcp://127.0.0.1:27665 >> abci_1.log 2>&1  &
-sleep 1
-sifnoded start --home ~/.sifnode-2 --p2p.laddr 0.0.0.0:27656  --grpc.address 0.0.0.0:9091 --address tcp://0.0.0.0:27660 --rpc.laddr tcp://127.0.0.1:27666 >> abci_2.log 2>&1  &
-sleep 1
-sifnoded start --home ~/.sifnode-3 --p2p.laddr 0.0.0.0:27657  --grpc.address 0.0.0.0:9092 --address tcp://0.0.0.0:27661 --rpc.laddr tcp://127.0.0.1:27667 >> abci_3.log 2>&1  &
-sleep 1
-
-rm -rf ~/.ibc-12/last-queried-heights.json
-rm -rf ~/.ibc-23/last-queried-heights.json
-rm -rf ~/.ibc-31/last-queried-heights.json
-rm -rf ~/.ibc-12/app.yaml
-rm -rf ~/.ibc-23/app.yaml
-rm -rf ~/.ibc-31/app.yaml
-printf "src: localnet-1\ndest: localnet-2\n" > ~/.ibc-12/app.yaml
-printf "src: localnet-2\ndest: localnet-3\n" > ~/.ibc-23/app.yaml
-printf "src: localnet-3\ndest: localnet-1\n" > ~/.ibc-31/app.yaml
-
-sleep 10
-ibc-setup ics20 --mnemonic "race draft rival universe maid cheese steel logic crowd fork comic easy truth drift tomorrow eye buddy head time cash swing swift midnight borrow" --home ~/.ibc-12
-sleep 1
-ibc-setup ics20 --mnemonic "race draft rival universe maid cheese steel logic crowd fork comic easy truth drift tomorrow eye buddy head time cash swing swift midnight borrow" --home ~/.ibc-23
-sleep 1
-ibc-setup ics20 --mnemonic "race draft rival universe maid cheese steel logic crowd fork comic easy truth drift tomorrow eye buddy head time cash swing swift midnight borrow" --home ~/.ibc-31
-
+echo "Starting sifnoded's"
 
 sleep 1
-echo "race draft rival universe maid cheese steel logic crowd fork comic easy truth drift tomorrow eye buddy head time cash swing swift midnight borrow" | ibc-relayer start -i -v --poll 10 --home ~/.ibc-12 >> ibc_12.log &
+sifnoded start --home ~/.sifnode-1 --p2p.laddr 0.0.0.0:27655  --grpc.address 0.0.0.0:9090 --grpc-web.address 0.0.0.0:9093 --address tcp://0.0.0.0:27659 --rpc.laddr tcp://127.0.0.1:27665 >> abci_1.log 2>&1  &
 sleep 1
-echo "race draft rival universe maid cheese steel logic crowd fork comic easy truth drift tomorrow eye buddy head time cash swing swift midnight borrow" | ibc-relayer start -i -v --poll 10 --home ~/.ibc-23 >> ibc_23.log &
+sifnoded start --home ~/.sifnode-2 --p2p.laddr 0.0.0.0:27656  --grpc.address 0.0.0.0:9091 --grpc-web.address 0.0.0.0:9094 --address tcp://0.0.0.0:27660 --rpc.laddr tcp://127.0.0.1:27666 >> abci_2.log 2>&1  &
 sleep 1
-echo "race draft rival universe maid cheese steel logic crowd fork comic easy truth drift tomorrow eye buddy head time cash swing swift midnight borrow" | ibc-relayer start -i -v --poll 10 --home ~/.ibc-31 >> ibc_31.log &
+sifnoded start --home ~/.sifnode-3 --p2p.laddr 0.0.0.0:27657  --grpc.address 0.0.0.0:9092 --grpc-web.address 0.0.0.0:9095 --address tcp://0.0.0.0:27661 --rpc.laddr tcp://127.0.0.1:27667 >> abci_3.log 2>&1 &
+sleep 5
+
+echo "Setting hermes"
+# copy hermes config to the hermes directory
+mkdir ~/.hermes
+cp scripts/hermes_config.toml ~/.hermes/config.toml
+
+hermes keys restore -m "race draft rival universe maid cheese steel logic crowd fork comic easy truth drift tomorrow eye buddy head time cash swing swift midnight borrow" localnet-1 --name sif
+hermes keys restore -m "race draft rival universe maid cheese steel logic crowd fork comic easy truth drift tomorrow eye buddy head time cash swing swift midnight borrow" localnet-2 --name sif
+hermes keys restore -m "race draft rival universe maid cheese steel logic crowd fork comic easy truth drift tomorrow eye buddy head time cash swing swift midnight borrow" localnet-3 --name sif
+
+# create hermes channels
+echo "Creating localnet-1 to localnet-2"
+hermes create channel localnet-1 localnet-2 --port-a transfer --port-b transfer -o unordered
+sleep 1
+echo "Creating localnet-2 to localnet-3"
+hermes create channel localnet-2 localnet-3 --port-a transfer --port-b transfer -o unordered
+sleep 1
+echo "Creating localnet-1 to localnet-3"
+hermes create channel localnet-1 localnet-3 --port-a transfer --port-b transfer -o unordered
+sleep 1
+
+# start hermes
+hermes start > hermes.log 2>&1 &
